requirements.txt
langchain
langchain_community
streamlit
pycryptodome
python-dotenv
tiktoken
unstructured
python-pptx
docx2txt
PyPDF
PyPDF2
openai==0.28.1
chromadb==0.3.29
azure-search-documents==11.4.0b8
azure-identity
xlrd
flask
fpdf==1.7.2
unstructured[xlsx]
unstructured[pdf]
python-docx
pdfkit
wkhtmltopdf
weasyprint
python-dateutil==2.8.2
numpy==1.24.2
scikit-learn==1.2.1
matplotlib==3.6.3
seaborn==0.12.2
pandas==1.5.3